// zoommeeting-angular/src/app/models/booth.model.ts
export interface boothInfo {

    // name:string
    // start:Time
    // end:Time
    // date:Date;
    boothid: string;
    company_name: string;
    company_video: string;
    company_brochure: string;
    company_email: string;
    is_available_for_one_on_one: boolean;
    company_meet: string;
}